+++ CHANGES.md
@@ -10,6 +10,7 @@ Beta Releases
    * Renamed `Simon1994PlanetaryPositions` functions `ComputeSunPositionInEarthInertialFrame` and `ComputeMoonPositionInEarthInertialFrame` to `computeSunPositionInEarthInertialFrame` and `computeMoonPositionInEarthInertialFrame`, respectively.
    * Replaced `Scene.scene2D.projection` property with read-only `Scene.mapProjection`.  Set this with the `mapProjection` option for the `Viewer`, `CesiumWidget`, or `Scene` constructors.
    * `Scene` constructor function now takes an `options` parameter instead of individual parameters.
+   * Replaced `color`, `outlineColor`, and `outlineWidth` in `DynamicPath` with a `material` property.
 * Added `Cesium.VERSION` to the combined `Cesium.js` file.
 * Fixed support for embedded resources in glTF models.
 * Added 'PolylineGlowMaterialProperty' which enables data sources to use the PolylineGlow material.